refactor(types): extract MetadataMdoPredicates + убрать dangling javadoc
